feat: review Docker configuration, rename some services and do good
