+++ .env.example
@@ -8,7 +8,7 @@ MEDIA_URL=/media/
 DB_NAME=databus
 DB_USER=postgres
 DB_PASSWORD=postgres
-DB_HOST=store
+DB_HOST=database
 DB_PORT=5432
 
 # RabbitMQ Configuration
@@ -33,7 +33,7 @@ PREFECT_API_URL=http://localhost:4200/api
 
 # Host Machine Port Configuration
 BACKEND_PORT=8000
-STORE_PORT=5432
+DATABASE_PORT=5432
 STATE_PORT=6379
 MQTT_BROKER_PORT=1883
 MESSAGE_BROKER_AMQP_PORT=5672
