+++ .gitignore
@@ -77,6 +77,7 @@ local_settings.py
 db.sqlite3
 db.sqlite3-journal
 staticfiles/
+migrations/
 
 # Flask stuff:
 instance/
