+++ backend/schedule_engine/aux_files/mbta_trip_update.json
