+++ backend/schedule_engine/tasks.py
@@ -41,7 +41,7 @@ def get_current_timestamp():
     return int(datetime.now().timestamp())
 
 
-@shared_task(queue="periodic_engine")
+@shared_task(queue="schedule_engine")
 def build_vehicle_positions():
     """
     Build the VehiclePosition feed message."""
@@ -130,19 +130,19 @@ def build_vehicle_positions():
 
     # Create and save JSON
     feed_message_json = json.dumps(feed_message)
-    with open("periodic_engine/files/vehicle_positions.json", "w") as f:
+    with open("schedule_engine/files/vehicle_positions.json", "w") as f:
         f.write(feed_message_json)
 
     # Create and save Protobuf
     feed_message_json = json.loads(feed_message_json)
     feed_message_pb = json_format.ParseDict(feed_message_json, gtfs_rt.FeedMessage())
-    with open("periodic_engine/files/vehicle_positions.pb", "wb") as f:
+    with open("schedule_engine/files/vehicle_positions.pb", "wb") as f:
         f.write(feed_message_pb.SerializeToString())
 
     return "FeedMessage VehiclePosition built successfully"
 
 
-@shared_task(queue="periodic_engine")
+@shared_task(queue="schedule_engine")
 def build_trip_updates():
     r = get_redis()
 
@@ -216,13 +216,13 @@ def build_trip_updates():
 
     # Create and save JSON
     feed_message_json = json.dumps(feed_message)
-    with open("periodic_engine/files/trip_updates.json", "w") as f:
+    with open("schedule_engine/files/trip_updates.json", "w") as f:
         f.write(feed_message_json)
 
     # Create and save Protobuf
     feed_message_json = json.loads(feed_message_json)
     feed_message_pb = json_format.ParseDict(feed_message_json, gtfs_rt.FeedMessage())
-    with open("periodic_engine/files/trip_updates.pb", "wb") as f:
+    with open("schedule_engine/files/trip_updates.pb", "wb") as f:
         f.write(feed_message_pb.SerializeToString())
 
     # Send status update to WebSocket
@@ -241,7 +241,7 @@ def build_trip_updates():
     return "Feed TripUpdate built."
 
 
-@shared_task(queue="periodic_engine")
+@shared_task(queue="schedule_engine")
 def build_alerts():
     print("Building feed Alert...")
     return "Feed ServiceAlert built"
