+++ backend/schedule_engine/views.py
@@ -11,19 +11,19 @@ def status(request):
 
 
 def schedule(request):
-    file_path = settings.BASE_DIR / "periodic_engine" / "files" / "bUCR_GTFS.zip"
+    file_path = settings.BASE_DIR / "schedule_engine" / "files" / "bUCR_GTFS.zip"
     return FileResponse(open(file_path, "rb"), filename="bUCR_GTFS.zip")
 
 
 @xframe_options_exempt
 def vehicle_json(request):
-    file_path = settings.BASE_DIR / "periodic_engine" / "files" / "vehicle_positions.json"
+    file_path = settings.BASE_DIR / "schedule_engine" / "files" / "vehicle_positions.json"
     return FileResponse(open(file_path, "rb"), filename="vehicle_positions.json")
 
 
 @xframe_options_exempt
 def vehicle_pb(request):
-    file_path = settings.BASE_DIR / "periodic_engine" / "files" / "vehicle_positions.pb"
+    file_path = settings.BASE_DIR / "schedule_engine" / "files" / "vehicle_positions.pb"
     return FileResponse(
         open(file_path, "rb"), as_attachment=True, filename="vehicle_positions.pb"
     )
@@ -31,13 +31,13 @@ def vehicle_pb(request):
 
 @xframe_options_exempt
 def trip_updates_json(request):
-    file_path = settings.BASE_DIR / "periodic_engine" / "files" / "trip_updates.json"
+    file_path = settings.BASE_DIR / "schedule_engine" / "files" / "trip_updates.json"
     return FileResponse(open(file_path, "rb"), filename="trip_updates.json")
 
 
 @xframe_options_exempt
 def trip_updates_pb(request):
-    file_path = settings.BASE_DIR / "periodic_engine" / "files" / "trip_updates.pb"
+    file_path = settings.BASE_DIR / "schedule_engine" / "files" / "trip_updates.pb"
     return FileResponse(
         open(file_path, "rb"), as_attachment=True, filename="trip_updates.pb"
     )
