+++ compose.dev.yml
@@ -17,7 +17,7 @@ services:
       - ./backend:/app
       - backend_venv:/home/app/.venv
     depends_on:
-      store:
+      database:
         condition: service_healthy
       state:
         condition: service_healthy
@@ -36,17 +36,17 @@ services:
       - ./backend:/app
       - backend_venv:/home/app/.venv
     depends_on:
-      store:
+      database:
         condition: service_healthy
       state:
         condition: service_healthy
       message-broker:
         condition: service_healthy
 
-  periodic-engine:
+  schedule-engine:
     build:
       context: ./backend
-      target: periodic-engine
+      target: schedule-engine
     environment:
       - DJANGO_SETUP=False
       - UV_PROJECT_ENVIRONMENT=/home/app/.venv
@@ -57,7 +57,7 @@ services:
       - ./backend:/app
       - backend_venv:/home/app/.venv
     depends_on:
-      store:
+      database:
         condition: service_healthy
       state:
         condition: service_healthy
@@ -78,22 +78,22 @@ services:
       - ./backend:/app
       - backend_venv:/home/app/.venv
     depends_on:
-      store:
+      database:
         condition: service_healthy
       state:
         condition: service_healthy
       message-broker:
         condition: service_healthy
 
-  store:
+  database:
     build:
-      context: ./store
+      context: ./database
     environment:
       - POSTGRES_DB=${DB_NAME:-databus}
       - POSTGRES_USER=${DB_USER:-postgres}
       - POSTGRES_HOST_AUTH_METHOD=trust
     volumes:
-      - store_data:/var/lib/postgresql
+      - database_data:/var/lib/postgresql
     healthcheck:
       test: ["CMD", "pg_isready", "-U", "${DB_USER:-postgres}", "-d", "${DB_NAME:-databus}"]
       interval: 10s
@@ -150,7 +150,7 @@ services:
     command: prefect server start
 
 volumes:
-  store_data:
+  database_data:
   state_data:
   lake_data:
   message_broker_data:
